fix: #271 - 휴식시간 UI 오차 수정
타입 변경, round 처리

+++ HowManySetWidget/Model/LiveActivityModel.swift
@@ -19,7 +19,7 @@ public struct WorkoutDataForLiveActivity: Equatable, Codable, Hashable {
     var currentRoutineCompleted: Bool
 
     var restStartDate: Date?
-    var restTime: Int 
+    var restTime: Float
     var isResting: Bool
     var isRestPaused: Bool
